[Add]PlayerInformationを追加
