+++ src/main/scala/com/github/unchama/seichiassist/util/Util.scala
@@ -3,7 +3,7 @@ package com.github.unchama.seichiassist.util
 import org.bukkit.ChatColor._
 import org.bukkit._
 import org.bukkit.block.{Block, Skull}
-import org.bukkit.entity.{LivingEntity, Player}
+import org.bukkit.entity.LivingEntity
 import org.bukkit.event.entity.EntityDamageEvent.DamageCause
 import org.bukkit.inventory.ItemStack
 import org.bukkit.inventory.meta.SkullMeta
@@ -41,19 +41,6 @@ object Util {
       line.contains("所有者：") && line.drop(line.indexOf("所有者：") + 4).toLowerCase == name
         .toLowerCase()
     )
-  }
-
-  def getPlayerDirection(player: Player): AbsoluteDirection = {
-    var rotation = ((player.getLocation.getYaw + 180) % 360).toDouble
-
-    if (rotation < 0) rotation += 360.0
-
-    // 0,360:south 90:west 180:north 270:east
-    if (0.0 <= rotation && rotation < 45.0) AbsoluteDirection.NORTH
-    else if (45.0 <= rotation && rotation < 135.0) AbsoluteDirection.EAST
-    else if (135.0 <= rotation && rotation < 225.0) AbsoluteDirection.SOUTH
-    else if (225.0 <= rotation && rotation < 315.0) AbsoluteDirection.WEST
-    else AbsoluteDirection.NORTH
   }
 
   def isVotingFairyPeriod(start: Calendar, end: Calendar): Boolean = {
